feat: use newtx instead of times
This change harmonizes the appearance of sans serif, serif, typewriter,
and math fonts throughout the document. It also means that he full set
of math symbols from mathtools and amssymb are provided, which means
that these packages no longer need to be loaded by users. The amsthms
package is loaded through newtx via an option.